# docs/openapi/order-api.yaml
openapi: 3.0.3
info:
  title: OMEcommerce — Order Management API
  version: "15.14.0"
  description: >
    The full order-management surface (plan/15): cart/checkout, admin order
    operations (fulfillment, refund, cancel, close, invoicing, shipment
    tracking, email), and the customer-facing storefront order routes. This
    is the first OpenAPI document in the repo — scoped to the order module,
    not the whole backend (see plan/15 Phase 14). Written by hand against the
    live route/schema definitions in src/modules/order and
    src/modules/customer, not auto-generated from the Zod schemas.
servers:
  - url: http://localhost:4100
    description: Local dev (scripts/dev-up.sh)
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT from POST /admin/v1/auth/login, carries RBAC permission codes.
    customerBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Customer JWT from POST /store/v1/customers/actions/login.
  schemas:
...
    Money:
      type: string
      description: Fixed-point decimal string, scale 4 (e.g. "45.0000").
      example: "45.0000"
    OrderLine:
      type: object
      properties:
        sku: { type: string }
        name: { type: string }
        qty: { type: integer }
        unitPrice: { $ref: '#/components/schemas/Money' }
        taxAmount: { $ref: '#/components/schemas/Money' }
        discountAmount: { $ref: '#/components/schemas/Money' }
        rowTotal: { $ref: '#/components/schemas/Money' }
        fulfilledQty: { type: integer }
        refundedQty: { type: integer }
    OrderAddress:
      type: object
      properties:
        type: { type: string, enum: [BILLING, SHIPPING] }
        name: { type: string }
        company: { type: string, nullable: true }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        region: { type: string, nullable: true }
        postalCode: { type: string }
        country: { type: string }
        phone: { type: string, nullable: true }
    PaymentTransaction:
      type: object
      properties:
        method: { type: string }
        gateway: { type: string }
        type: { type: string, enum: [AUTHORIZE, CAPTURE, REFUND, VOID] }
        amount: { $ref: '#/components/schemas/Money' }
        currency: { type: string }
        status: { type: string, enum: [PENDING, SUCCEEDED, FAILED] }
        gatewayRef: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    Fulfillment:
      type: object
      properties:
        publicId: { type: string, format: uuid }
        status: { type: string, enum: [PENDING, PACKED, SHIPPED, DELIVERED, CANCELLED] }
        trackingNumber: { type: string, nullable: true }
        carrier: { type: string, nullable: true }
        carrierTrackingUrl: { type: string, nullable: true }
        estimatedDeliveryAt: { type: string, format: date-time, nullable: true }
        currentStatus: { type: string, nullable: true }
        shippingNotes: { type: string, nullable: true }
        hasPackingSlip: { type: boolean }
        shippedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        lines:
          type: array
          items:
            type: object
            properties: { sku: { type: string }, qty: { type: integer } }
    OrderInvoice:
      type: object
      properties:
        publicId: { type: string, format: uuid }
        invoiceNumber: { type: string }
        status: { type: string, enum: [DRAFT, ISSUED] }
        subtotal: { $ref: '#/components/schemas/Money' }
        discountTotal: { $ref: '#/components/schemas/Money' }
        taxTotal: { $ref: '#/components/schemas/Money' }
        grandTotal: { $ref: '#/components/schemas/Money' }
        createdAt: { type: string, format: date-time }
        lines:
          type: array
          items:
            type: object
            properties:
              sku: { type: string }
              qty: { type: integer }
              unitPrice: { $ref: '#/components/schemas/Money' }
              taxAmount: { $ref: '#/components/schemas/Money' }
              rowTotal: { $ref: '#/components/schemas/Money' }
    OrderNote:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [INTERNAL, CUSTOMER] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    OrderView:
      type: object
      properties:
        publicId: { type: string, format: uuid }
        orderNumber: { type: string }
        email: { type: string, format: email }
        currency: { type: string }
        status: { type: string, enum: [PENDING, CONFIRMED, PROCESSING, ON_HOLD, COMPLETED, CLOSED, CANCELLED] }
        financialStatus: { type: string, enum: [PENDING, AUTHORIZED, PARTIALLY_PAID, PAID, PARTIALLY_REFUNDED, REFUNDED, VOIDED, FAILED] }
        fulfillmentStatus: { type: string, enum: [UNFULFILLED, PARTIALLY_FULFILLED, FULFILLED, RETURNED] }
        subtotal: { $ref: '#/components/schemas/Money' }
        discountTotal: { $ref: '#/components/schemas/Money' }
        taxTotal: { $ref: '#/components/schemas/Money' }
        shippingTotal: { $ref: '#/components/schemas/Money' }
        grandTotal: { $ref: '#/components/schemas/Money' }
        shippingMethodCode: { type: string, nullable: true }
        customerIp: { type: string, nullable: true }
        placedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time, nullable: true }
        lines: { type: array, items: { $ref: '#/components/schemas/OrderLine' } }
        addresses: { type: array, items: { $ref: '#/components/schemas/OrderAddress' } }
        payments: { type: array, items: { $ref: '#/components/schemas/PaymentTransaction' } }
        fulfillments: { type: array, items: { $ref: '#/components/schemas/Fulfillment' } }
        invoices: { type: array, items: { $ref: '#/components/schemas/OrderInvoice' } }
        notes: { type: array, items: { $ref: '#/components/schemas/OrderNote' } }
...
paths:
...
  /store/v1/carts/{publicId}/checkout:
    post:
      summary: Complete checkout (cart -> reserve -> order -> capture payment)
      tags: [Storefront Checkout]
      parameters: [{ name: publicId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email, billingAddress, shippingAddress, shippingMethodCode, paymentMethod]
              properties:
                email: { type: string, format: email }
                shippingMethodCode: { type: string }
                paymentMethod: { type: string }
                testScenario: { type: string, enum: [approve, decline] }
      responses:
        '201': { description: Created, content: { application/json: { schema: { type: object, properties: { data: { $ref: '#/components/schemas/OrderView' } } } } } }
        '402': { description: Payment declined }
        '409': { description: "Cart not active, or insufficient stock" }
...